requirements.txt
flask
google-genai
python-dotenv
pypdf
gunicorn